+++ avisaf/util/vectorizers.py
